Collapse ConnectionWidgetBase out of hierarchy
 - Fixes #843
 - Remove ConnectionWidgetBase class
 - Migrate most functionality into SetupStepComPortManual
 - Derive SetupConnectionWidgetBase from WizardPanel instead of CWB
