+++ PrinterControls/PrinterConnections/BaseConnectionWidget.cs
@@ -230,133 +230,4 @@ namespace MatterHackers.MatterControl.PrinterControls.PrinterConnections
 			base.OnDraw(graphics2D);
 		}
 	}
-
-	public class ConnectionWidgetBase : WizardPanel
-	{
-		private static Regex linuxDefaultUIFilter = new Regex("/dev/ttyS*\\d+", RegexOptions.CultureInvariant | RegexOptions.Compiled);
-		protected List<SerialPortIndexRadioButton> SerialPortButtonsList = new List<SerialPortIndexRadioButton>();
-		private bool printerComPortIsAvailable = false;
-
-		private event EventHandler unregisterEvents;
-
-		private PrinterInfo activePrinter = null;
-		protected PrinterInfo ActivePrinter
-		{
-			get
-			{
-				if(activePrinter == null)
-				{
-					var settings = ActiveSliceSettings.Instance;
-					activePrinter = new PrinterInfo
-					{
-						AutoConnect = settings.DoAutoConnect(),
-						BaudRate = settings.BaudRate(),
-						ComPort = settings.ComPort(),
-						DriverType = settings.DriverType(),
-						Id = settings.ID,
-						Name = settings.Name()
-					};
-				}
-
-				return activePrinter;
-			}
-		}
-
-		public ConnectionWidgetBase(WizardWindow wizard)
-			: base(wizard)
-		{
-			ActiveTheme.Instance.ThemeChanged.RegisterEvent((s,e) => this.Invalidate(), ref unregisterEvents);
-		}
-
-		public int GetPrinterRecordCount()
-		{
-			return Datastore.Instance.RecordCount("Printer");
-		}
-
-		public override void OnClosed(EventArgs e)
-		{
-			unregisterEvents?.Invoke(this, null);
-			base.OnClosed(e);
-		}
-
-		private SerialPortIndexRadioButton createComPortOption(string portName, bool isActivePrinterPort)
-		{
-			SerialPortIndexRadioButton comPortOption = new SerialPortIndexRadioButton(portName, portName)
-			{
-				HAnchor = HAnchor.ParentLeft,
-				Margin = new BorderDouble(3, 3, 5, 3),
-				TextColor = ActiveTheme.Instance.PrimaryTextColor,
-				Checked = isActivePrinterPort
-			};
-			return comPortOption;
-		}
-
-		protected string GetSelectedSerialPort()
-		{
-			foreach (SerialPortIndexRadioButton button in SerialPortButtonsList)
-			{
-				if (button.Checked)
-				{
-					return button.PortValue;
-				}
-			}
-
-			throw new Exception(LocalizedString.Get("Could not find a selected button."));
-		}
-
-		protected void CreateSerialPortControls(FlowLayoutWidget comPortContainer, string activePrinterSerialPort)
-		{
-			int portIndex = 0;
-			string[] allPorts = FrostedSerialPort.GetPortNames();
-			IEnumerable<string> filteredPorts;
-
-			if (OsInformation.OperatingSystem == OSType.X11)
-			{
-				// A default and naive filter that works well on Ubuntu 14
-				filteredPorts = allPorts.Where(portName => portName != "/dev/tty" && !linuxDefaultUIFilter.Match(portName).Success);
-			}
-			else
-			{
-				// looks_like_mac -- serialPort.StartsWith("/dev/tty."); looks_like_pc -- serialPort.StartsWith("COM")
-				filteredPorts = allPorts.Where(portName => portName.StartsWith("/dev/tty.") || portName.StartsWith("COM"));
-			}
-
-			IEnumerable<string> portsToCreate = filteredPorts.Any() ? filteredPorts : allPorts;
-
-			// Add a radio button for each filtered port
-			foreach (string portName in portsToCreate)
-			{
-				SerialPortIndexRadioButton comPortOption = createComPortOption(portName, activePrinterSerialPort == portName);
-				if (comPortOption.Checked)
-				{
-					printerComPortIsAvailable = true;
-				}
-
-				SerialPortButtonsList.Add(comPortOption);
-				comPortContainer.AddChild(comPortOption);
-
-				portIndex++;
-			}
-
-			// Add a virtual entry for serial ports that were previously configured but are not currently connected
-			if (!printerComPortIsAvailable && activePrinterSerialPort != null)
-			{
-				SerialPortIndexRadioButton comPortOption = createComPortOption(activePrinterSerialPort, true);
-				comPortOption.Enabled = false;
-
-				comPortContainer.AddChild(comPortOption);
-				SerialPortButtonsList.Add(comPortOption);
-				portIndex++;
-			}
-
-			//If there are still no com ports show a message to that effect
-			if (portIndex == 0)
-			{
-				TextWidget comPortOption = new TextWidget(LocalizedString.Get("No COM ports available"));
-				comPortOption.Margin = new BorderDouble(3, 6, 5, 6);
-				comPortOption.TextColor = ActiveTheme.Instance.PrimaryTextColor;
-				comPortContainer.AddChild(comPortOption);
-			}
-		}
-	}
 }
